Fix `literal_only_boolean_expressions` `is` False-Positive
Prior to this commit - a conditonal expression which used `is` to test against a type
parameter was treated as literal expression causing the lint to flag such conditions.

The actual type of a type parameter is not known for a program until a literal type is
substituted hence the flagging was a false-positive. This commit prevents such flagging.

Change-Id: I7b33510e83354fb652b159df452ed5063f51e875
Bug: #61903
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461680
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Auto-Submit: Rohit Saily <hello@rohitsaily.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>