[stable] [CFE/kernel] Clear invalidated ClassHierarchySubtypes
When recompiling (`computeDelta`) the class hierarchy is made up
to date and the old ClassHierarchySubtypes (if any) is marked as
invalid, but was erroneously not cleared, meaning that when trying
to use it (via expression compilation) it would through.
This fixes it be clearing it, meaning that it will be recreated in a
correct state if/when needed.

Fixes https://github.com/flutter/flutter/issues/178740

Risk: Low

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/464380
Change-Id: Ib2dbca09c4d483fa9d798c1d147c164737e73aff
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/464860
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>