[stable][analyzer] Dot shorthands: Error when using type arguments on a factory constructor in an abstract class.
Fixing a small bug where we wouldn't check the other error cases if we
had an abstract class as the context type.

Added a few language tests and a unit test for this case.

Issue description: When attempting to use type arguments on a dot
shorthand constructor which is invoking a factory constructor in an
abstract class, the analyzer isn't reporting an error.

What is the fix: Change the code flow so that we are checking for type
arguments when resolving to a factory constructor in an abstract class.

Why cherry-pick: Users won't have a static error for this and will need
to compile in order to find that their code is incorrect.

Risk: Low, this fix has landed on the main channel and is tested with language tests and unit tests. Extra tests covering the changes with https://dart-review.googlesource.com/c/sdk/+/463466 and https://dart-review.googlesource.com/c/sdk/+/464764

Issue link(s): https://github.com/dart-lang/sdk/issues/61978

Bug: https://github.com/dart-lang/sdk/issues/61978
Change-Id: Ic36f3159f2cc6b1f87fe22e21f72a7124095019d
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/462542
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463502
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Leaf Petersen <leafp@google.com>