[vm,dyn_modules] Fix evaluation expression function self-reference.
If the evaluation expression function is for an expression in the scope
of a generic function with a parameter whose type includes one of its
type parameters, then the graph for the parameter type will contain a
self reference to the evaluation expression function.

Normally, this self reference would be resolved via the scoped_function_
field. However, while scoped_function_class_ is set to the class containing the expression evaluation function (as expected),
if ReadObject() encounters that class, it returns the "real" class
the evaluation expression should be evaluated in. This means the
self-reference isn't detected.

Add an additional check for the right name/class for the expression
evaluation function.

TEST=pkg/vm_service/test/evaluate_function_type_parameters_test

Fixes: https://github.com/dart-lang/sdk/issues/61948
Cq-Include-Trybots: luci.dart.try:vm-linux-release-x64-try,vm-dyn-linux-debug-x64-try,vm-dyn-mac-debug-arm64-try
Change-Id: Id04735059d00d964a8f00885857d3255c18ecf99
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/464520
Commit-Queue: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Auto-Submit: Tess Strickland <sstrickl@google.com>