DeCo. Stop using isSynthetic in index.
There is no change in behavior. Index record relations between elements
and locations, where the element is identified by a triplet of names
(topLevelName, classMemberName, formalParameterName). Which is almost
enough, but for field / getter / setter the names are the same, so we
add the kind. It was more complicated than necessary, so I simplified
it a little.

Change-Id: If91b8881f89ffc7a8f41331ded9ef4b1b64136b8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/465965
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>