[analysis_server] Prevent LSP format-on-type from formatting in unwanted cases
We register `}` and `;` as trigger characters for format-on-type, but some instances of these characters should not trigger formatting (for example in strings and comments).

With this change, we'll look at the AST and only format for specific cases we've opted in to, instead of all.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5832

Change-Id: I94d348520649ea2e072f13350863f27a2e07699d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466100
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>