[messages] Ignore case in LinterRuleOptionsValidator.
Now that lint name matching is case insensitive
(https://dart-review.googlesource.com/c/sdk/+/465964), it makes sense
for the LinterRuleOptionsValidator to ignore case when checking the
contents of `analysis_options.yaml` files.

This paves the way for a follow-up CL that will change the names
declared in the linter to all lower case.


Change-Id: I6a6a696430f1d99408902b115587ac8c3d22fbe8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466000
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>