[messages] Prepare to standardize DiagnosticCode names to all lower case.
In a follow-up CL, I will change the code generator for diagnostic
messages so that all code-generated instances of `DiagnosticCode` have
`name` and `uniqueName` fields in all lower case letters. This will
help pave the way toward resolving conflicts between the diagnostic
code naming conventions between the analyzer and CFE.

To prepare for this change, the following pieces of code need to first
be modified to be case insensitive:

- The `_ignoreNullSafetyWarnings` check in
  `RemoveDeadCodeSoundFlowAnalysisTest`.

- The `_diagnosticCodes` fields in `_CannotIgnoreOptionValidator` and
  `_ErrorFilterOptionValidator`.

- The `computeErrorData` methods in `ConstantsDataComputer` and
  `_InheritanceDataComputer`.

- The `_writeDiagnostic` methods in `DriverEventsPrinter` and
  `ResolvedUnitResultPrinter`.

- The `_validateMessages` and `_validateSnippet` methods in
  `DocumentationValidator`.

- The `run` method in `TextualOutline`.

In most cases, the code was changed to be case insensitive by having
it call `.toLowerCase()` or `.toUpperCase()` on the diagnostic code
name before operating on it. In one case I was able to re-use the
`AnalyzerCode.lowerSnakeCaseName` getter.

Change-Id: I6a6a6964f3cc4c71f2e82478855d64d5ef91d26b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/465804
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>