[messages] Use lower case diagnostic names in _fe_analyzer_shared.
Changes the logic in `pkg/_fe_analyzer_shared` to use
`DiagnosticCode.lowerCaseName` instead of `DiagnosticCode.name`, and
`DiagnosticCode.lowerCaseUniqueName` instead of
`DiagnosticCode.uniqueName`. This ensures that diagnostic codes are
matched in a case-insensitive fashion.

This paves the way for deprecating (and eventually removing) the
`DiagnosticCode.name` and `DiagnosticCode.uniqueName` getters.

Change-Id: I6a6a6964528a215173e14131b2425896b5efa5d2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466183
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>