[messages] Use lower case analyzer diagnostic names in front_end.
Changes the logic in `pkg/front_end` to use
`DiagnosticCode.lowerCaseName` instead of `DiagnosticCode.name`. This
ensures that diagnostic codes are matched in a case-insensitive
fashion.

Note that this doesn't affect the behavior of the front end, just the
behavior of front end tests that invoke the analyzer.

This paves the way for deprecating (and eventually removing) the
`DiagnosticCode.name` getter.

Change-Id: I6a6a696459c9170ccf5b0c42622cfe3415696e98
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466282
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>