Fix when the warning about methods named factory is produced
Based on the conversation in
  https://github.com/dart-lang/language/issues/4586
it appears that we might allow both `augment` and `external` to apply
to the same method. While we are not currently supporting the
`augmentations` language feature, this CL will guard against the
possibility that `augmentations` might be enabled before the
`primary-constructors` feature.

Change-Id: I8a2133385c6afc919a814c38929d4c451631c49a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466841
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>