[messages] Fix correction message for replacedLint diagnostic.
Fixes the correction message for the `replacedLint`
diagnostic. Previously, the correction message referred to the wrong
parameter, so a problem message like `'lint_foo' was replaced by
'lint_bar' in Dart '1.2.3'` would be accompanied by the correction
message `Replace 'lint_foo' with '1.2.3'.`. Clearly the intention was
for the message to say `Replace 'lint_foo' with 'lint_bar'.`.

I'm currently working on a set of CLs to update all the messages in
the analyzer's `messages.yaml` file, replacing the placeholder
parameters `p0`, `p1`, `p2`, etc. with meaningful names. In the
process I discovered this mistake, so to ensure that the other CL's
don't change behavior, I've isolated the work for the `replacedLint`
diagnostic to its own CL. I've gone ahead and included renaming the
parameters as part of this CL, since that makes it much easier to see
that the fix is correct.

Change-Id: I6a6a696455ffa13b90510ec8a6052b6c23dce6d8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/467387
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>