analyzer: Refactor invalid_widget_preview tests to be more individualized
I need to modify these tests for a change that was made to the Preview
annotation, and these tests are very hard to reason about with so
many expected errors and such massive offsets. Each test should test
one thing, and this moves them in that direction.

Change-Id: Ifa3d11c1336966f6198ac4296a151f27bca842aa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/467443
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>