[analyzer] Dot shorthands: Remove error when a constructor conflicts with a non-static member.
Users are getting an extra error when a constructor and an instance method/getter have the same name.
In the implementation, we produce an error once we see that the retrieved dot shorthand element isn't static. Instead, we should be falling through to check for constructor elements of that name.

This CL makes the necessary changes.

Testing:
`element != null && element is InternalExecutableElement && element.isStatic` is tested by existing tests.

`element != null && element is InternalExecutableElement && !element.isStatic` is tested in the `test_conflict_instance_*` unit tests. This is the behavior with the bug.

`element != null && element is! InternalExecutableElement` should not
exist since the `_resolveElement only returns `ExecutableElement`s or
`null`. (There's definitely something we can refactor here, but I'll leave it for a future CL.)

If we can't find a static method/field of that name, we should be
checking if it's the name of a constructor for both of the above cases.

`element == null` occurs when we can't find any method/getter of a given
name. If there's a setter of the given name, this will be `true` as well. This is well tested by prior unit tests.

Fixes: https://github.com/dart-lang/sdk/issues/62045
Change-Id: Icc6b66a95cfc79312891453b6f277f1347e059bc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/467261
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>