[dart2wasm] Optimize fine grained partitioning for deferred loading
This saves around 730 KB (-7.7%) on essentials main module.

The existing fine grained partitioning of code into wasm modules
operates on static elements. That means as soon as a constructor is
called, all instance methods of a class are included as well.

This CL improves upon this using the insight that some instance methods
are always accessed via direct calls (i.e. all call sites to the
instance member are devirtualized).

For those methods that are never called indirectly, the partitioning
algorithm can avoid enquing those instance methods when the class is
enqueued and instead enqueue them when the devirtualized call site is
hit.

An example where this can happen is e.g. main module allocates a proto
class but doesn't access field X. Currently when hitting the constructor
we'll enqueue all instance members. Now we don't enqueue X (as all uses
of X are direct calls to the X getter). That means if X is only used in
a deferred module, then we can move the X getter to that deferred
module.

Change-Id: Ib45edb6e658a62aaeb15ed7c40745733ec619ff0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/467700
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>