[messages] Use lower case name for accessing customized codes.
Changes the logic for accessing the maps `customizedCorrections` and
`customizedMessages` so that it looks up messages by
`lowerCaseUniqueName` rather than `uniqueName`.

These maps are used by the google3 import of the analyzer, so that
error messages can be customized for internal use without having to
patch analyzer sources. This change is a no-op because all diagnostic
codes currently have lower case names.

This paves the way for deprecating (and eventually removing) the
`DiagnosticCode.uniqueName` getter.

Change-Id: I6a6a69646e7e131976ef3ac708f83b1e82b47ead
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/466128
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>